requirements.txt
PySide6>=6.8.0
numpy>=2.0.0
av>=14.0.0
opencv-python-headless>=4.10.0
imageio>=2.30.0
OpenEXR>=3.0.0
transnetv2-pytorch>=1.0.5
python-mpv>=1.0.0
